csgo_pros/requirements.txt
pymongo==3.10.1
python-dotenv==0.13.0
Scrapy~=2.1.0
numpy~=1.18.4
console-menu==0.6.0
scikit-learn~=0.23.1
matplotlib~=3.2.1
pandas~=1.0.3
scipy~=1.4.1
jupiter~=1.0.0